feat(native-participants-pane) participants pane open/close fixed
