+++ react/features/participants-pane/actions.any.js
@@ -1,0 +1,28 @@
+// @flow
+
+import {
+    PARTICIPANTS_PANE_CLOSE,
+    PARTICIPANTS_PANE_OPEN
+} from './actionTypes';
+
+/**
+ * Action to close the participants pane.
+ *
+ * @returns {Object}
+ */
+export const close = () => {
+    return {
+        type: PARTICIPANTS_PANE_CLOSE
+    };
+};
+
+/**
+ * Action to open the participants pane.
+ *
+ * @returns {Object}
+ */
+export const open = () => {
+    return {
+        type: PARTICIPANTS_PANE_OPEN
+    };
+};
