+++ react/features/participants-pane/actions.native.js
@@ -1,32 +1,9 @@
+// @flow
+
 import { openDialog } from '../base/dialog';
 
-import { PARTICIPANTS_PANE_CLOSE, PARTICIPANTS_PANE_OPEN } from './actionTypes';
 import { ContextMenuLobbyParticipantReject, ContextMenuMeetingParticipantDetails } from './components/native';
-
-
-/**
- * Action to open the participants pane.
- *
- * @returns {Object}
- */
-export function open() {
-    console.log(2);
-
-    return {
-        type: PARTICIPANTS_PANE_OPEN
-    };
-}
-
-/**
- * Action to close the participants pane.
- *
- * @returns {Object}
- */
-export function close() {
-    return {
-        type: PARTICIPANTS_PANE_CLOSE
-    };
-}
+export * from './actions.any';
 
 /**
  * Displays the context menu for the selected lobby participant.
@@ -34,7 +11,7 @@ export function close() {
  * @param {Object} participant - The selected lobby participant.
  * @returns {Function}
  */
-export function showContextMenuReject(participant) {
+export function showContextMenuReject(participant: Object) {
     return openDialog(ContextMenuLobbyParticipantReject, { participant });
 }
 
@@ -45,6 +22,6 @@ export function showContextMenuReject(participant) {
  * @param {Object} participant - The selected meeting participant.
  * @returns {Function}
  */
-export function showContextMenuDetails(participant) {
+export function showContextMenuDetails(participant: Object) {
     return openDialog(ContextMenuMeetingParticipantDetails, { participant });
 }
