+++ react/features/participants-pane/actions.web.js
@@ -1,26 +1,3 @@
-import {
-    PARTICIPANTS_PANE_CLOSE,
-    PARTICIPANTS_PANE_OPEN
-} from './actionTypes';
+// @flow
 
-/**
- * Action to close the participants pane.
- *
- * @returns {Object}
- */
-export const close = () => {
-    return {
-        type: PARTICIPANTS_PANE_CLOSE
-    };
-};
-
-/**
- * Action to open the participants pane.
- *
- * @returns {Object}
- */
-export const open = () => {
-    return {
-        type: PARTICIPANTS_PANE_OPEN
-    };
-};
+export * from './actions.any';
